Switch SDKLookup tests to use mock hostpolicy
These are just testing discovery of the SDK locations, and do not need
an actual app or the SDK to run.
