+++ src/test/HostActivationTests/FrameworkResolution/FrameworkResolutionBase.cs
@@ -142,72 +142,6 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation.FrameworkResolution
                     Directory.Delete(_baseDir, true);
                 }
             }
-
-            public class DotNetBuilder
-            {
-                private readonly string _path;
-                private readonly RepoDirectoriesProvider _repoDirectories;
-
-                public DotNetBuilder(string basePath, string builtDotnet, string name)
-                {
-                    _path = Path.Combine(basePath, name);
-                    Directory.CreateDirectory(_path);
-
-                    _repoDirectories = new RepoDirectoriesProvider(builtDotnet: _path);
-
-                    // Prepare the dotnet installation mock
-
-                    // ./dotnet.exe - used as a convenient way to load and invoke hostfxr. May change in the future to use test-specific executable
-                    var builtDotNetCli = new DotNetCli(builtDotnet);
-                    File.Copy(
-                        builtDotNetCli.DotnetExecutablePath,
-                        Path.Combine(_path, RuntimeInformationExtensions.GetExeFileNameForCurrentPlatform("dotnet")),
-                        true);
-
-                    // ./host/fxr/<version>/hostfxr.dll - this is the component being tested
-                    SharedFramework.CopyDirectory(
-                        builtDotNetCli.GreatestVersionHostFxrPath,
-                        Path.Combine(_path, "host", "fxr", Path.GetFileName(builtDotNetCli.GreatestVersionHostFxrPath)));
-                }
-
-                public DotNetBuilder AddMicrosoftNETCoreAppFramework(string version)
-                {
-                    // ./shared/Microsoft.NETCore.App/<version> - create a mock of the root framework
-                    string netCoreAppPath = Path.Combine(_path, "shared", MicrosoftNETCoreApp, version);
-                    Directory.CreateDirectory(netCoreAppPath);
-
-                    // ./shared/Microsoft.NETCore.App/<version>/hostpolicy.dll - this is a mock, will not actually load CoreCLR
-                    string mockHostPolicyFileName = RuntimeInformationExtensions.GetSharedLibraryFileNameForCurrentPlatform("mockhostpolicy");
-                    File.Copy(
-                        Path.Combine(_repoDirectories.Artifacts, "corehost_test", mockHostPolicyFileName),
-                        Path.Combine(netCoreAppPath, RuntimeInformationExtensions.GetSharedLibraryFileNameForCurrentPlatform("hostpolicy")),
-                        true);
-
-                    return this;
-                }
-
-                public DotNetBuilder AddFramework(
-                    string name,
-                    string version,
-                    Action<RuntimeConfig> runtimeConfigCustomizer)
-                {
-                    // ./shared/<name>/<version> - create a mock of effectively empty non-root framework
-                    string path = Path.Combine(_path, "shared", name, version);
-                    Directory.CreateDirectory(path);
-
-                    // ./shared/<name>/<version>/<name>.runtimeconfig.json - runtime config which can be customized
-                    RuntimeConfig runtimeConfig = new RuntimeConfig(Path.Combine(path, name + ".runtimeconfig.json"));
-                    runtimeConfigCustomizer(runtimeConfig);
-                    runtimeConfig.Save();
-
-                    return this;
-                }
-
-                public DotNetCli Build()
-                {
-                    return new DotNetCli(_path);
-                }
-            }
         }
     }
 }
