+++ src/test/HostActivationTests/MultilevelSDKLookup.cs
@@ -2,11 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.Win32;
+using Microsoft.DotNet.Cli.Build;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Runtime.InteropServices;
 using Xunit;
 
@@ -22,8 +21,8 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             {"COMPlus_ReadyToRun", "0" },
         };
 
-        private RepoDirectoriesProvider RepoDirectories;
-        private TestProjectFixture PortableAppFixture;
+        private readonly RepoDirectoriesProvider RepoDirectories;
+        private readonly DotNetCli DotNet;
 
         private readonly string _currentWorkingDir;
         private readonly string _userDir;
@@ -35,36 +34,29 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
         private readonly string _regSdkBaseDir;
         private readonly string _exeSelectedMessage;
         private readonly string _regSelectedMessage;
-        private readonly string _sdkDir;
         private readonly string _multilevelDir;
 
         private const string _dotnetSdkDllMessageTerminator = "dotnet.dll]";
 
         public MultilevelSDKLookup()
         {
-            // From the artifacts dir, it's possible to find where the sharedFrameworkPublish folder is. We need
-            // to locate it because we'll copy its contents into other folders
-            string artifactsDir = Environment.GetEnvironmentVariable("TEST_ARTIFACTS");
-            string builtDotnet = Path.Combine(artifactsDir, "sharedFrameworkPublish");
-
             // The dotnetMultilevelSDKLookup dir will contain some folders and files that will be
             // necessary to perform the tests
-            string baseMultilevelDir = Path.Combine(artifactsDir, "dotnetMultilevelSDKLookup");
+            string baseMultilevelDir = Path.Combine(TestArtifact.TestArtifactsPath, "dotnetMultilevelSDKLookup");
             _multilevelDir = SharedFramework.CalculateUniqueTestDirectory(baseMultilevelDir);
 
-            // The three tested locations will be the cwd, the user folder and the exe dir. cwd and user are no longer supported.
+            // The tested locations will be the cwd, user folder, exe dir, and registered directory. cwd and user are no longer supported.
             //     All dirs will be placed inside the multilevel folder
-
             _currentWorkingDir = Path.Combine(_multilevelDir, "cwd");
             _userDir = Path.Combine(_multilevelDir, "user");
             _exeDir = Path.Combine(_multilevelDir, "exe");
             _regDir = Path.Combine(_multilevelDir, "reg");
 
-            // It's necessary to copy the entire publish folder to the exe dir because
-            // we'll need to build from it. The CopyDirectory method automatically creates the dest dir
-            SharedFramework.CopyDirectory(builtDotnet, _exeDir);
+            DotNet = new DotNetBuilder(_multilevelDir, Path.Combine(TestArtifact.TestArtifactsPath, "sharedFrameworkPublish"), "exe")
+                .AddMicrosoftNETCoreAppFramework("9999.0.0")
+                .Build();
 
-            RepoDirectories = new RepoDirectoriesProvider(builtDotnet: _exeDir);
+            RepoDirectories = new RepoDirectoriesProvider(builtDotnet: DotNet.BinPath);
 
             // SdkBaseDirs contain all available version folders
             _cwdSdkBaseDir = Path.Combine(_currentWorkingDir, "sdk");
@@ -78,32 +70,6 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             Directory.CreateDirectory(_exeSdkBaseDir);
             Directory.CreateDirectory(_regSdkBaseDir);
 
-            // Restore and build PortableApp from exe dir
-            PortableAppFixture = new TestProjectFixture("PortableApp", RepoDirectories)
-                .EnsureRestored(RepoDirectories.CorehostPackages)
-                .BuildProject();
-            var fixture = PortableAppFixture;
-
-            // Set a dummy framework version (9999.0.0) in the exe sharedFx location. We will
-            // always pick the framework from this to avoid interference with the sharedFxLookup
-            string exeDirDummyFxVersion = Path.Combine(_exeDir, "shared", "Microsoft.NETCore.App", "9999.0.0");
-            string builtSharedFxDir = fixture.BuiltDotnet.GreatestVersionSharedFxPath;
-            SharedFramework.CopyDirectory(builtSharedFxDir, exeDirDummyFxVersion);
-
-            // The actual SDK version can be obtained from the built fixture. We'll use it to
-            // locate the sdkDir from which we can get the files contained in the version folder
-            string sdkBaseDir = Path.Combine(fixture.SdkDotnet.BinPath, "sdk");
-
-            var sdkVersionDirs = Directory.EnumerateDirectories(sdkBaseDir)
-                .Select(p => Path.GetFileName(p));
-
-            string greatestVersionSdk = sdkVersionDirs
-                .Where(p => !string.Equals(p, "NuGetFallbackFolder", StringComparison.OrdinalIgnoreCase))
-                .OrderByDescending(p => p.ToLower())
-                .First();
-
-            _sdkDir = Path.Combine(sdkBaseDir, greatestVersionSdk);
-
             // Trace messages used to identify from which folder the SDK was picked
             _exeSelectedMessage = $"Using dotnet SDK dll=[{_exeSdkBaseDir}";
             _regSelectedMessage = $"Using dotnet SDK dll=[{_regSdkBaseDir}";
@@ -111,9 +77,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
 
         public void Dispose()
         {
-            PortableAppFixture.Dispose();
-
-            if (!TestProject.PreserveTestRuns())
+            if (!TestArtifact.PreserveTestRuns())
             {
                 Directory.Delete(_multilevelDir, true);
             }
@@ -128,11 +92,6 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 return;
             }
 
-            var fixture = PortableAppFixture
-                .Copy();
-
-            var dotnet = fixture.BuiltDotnet;
-
             // Set specified SDK version = 9999.3.4-global-dummy
             SetGlobalJsonVersion("SingleDigit-global.json");
 
@@ -142,7 +101,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: empty
             // Reg: empty
             // Expected: no compatible version and a specific error messages
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -163,7 +122,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.4.1, 9999.3.4-dummy
             // Reg: empty
             // Expected: no compatible version and a specific error message
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -184,7 +143,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.4.1, 9999.3.4-dummy
             // Reg: 9999.3.3
             // Expected: no compatible version and a specific error message
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -205,7 +164,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.4.1, 9999.3.4-dummy, 9999.3.4
             // Reg: 9999.3.3
             // Expected: 9999.3.4 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -226,7 +185,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.4.1, 9999.3.4-dummy, 9999.3.4
             // Reg: 9999.3.3, 9999.3.5-dummy
             // Expected: 9999.3.5-dummy from reg dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -247,7 +206,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.4.1, 9999.3.4-dummy, 9999.3.4, 9999.3.600
             // Reg: 9999.3.3, 9999.3.5-dummy
             // Expected: 9999.3.5-dummy from reg dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -268,7 +227,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.4.1, 9999.3.4-dummy, 9999.3.4, 9999.3.600, 9999.3.4-global-dummy
             // Reg: 9999.3.3, 9999.3.5-dummy
             // Expected: 9999.3.4-global-dummy from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -281,7 +240,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 .And.HaveStdErrContaining(Path.Combine(_exeSelectedMessage, "9999.3.4-global-dummy", _dotnetSdkDllMessageTerminator));
 
             // Verify we have the expected SDK versions
-            dotnet.Exec("--list-sdks")
+            DotNet.Exec("--list-sdks")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -308,11 +267,6 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 return;
             }
 
-            var fixture = PortableAppFixture
-                .Copy();
-
-            var dotnet = fixture.BuiltDotnet;
-
             // Set specified SDK version = 9999.3.304-global-dummy
             SetGlobalJsonVersion("TwoPart-global.json");
 
@@ -322,7 +276,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: empty
             // Reg: empty
             // Expected: no compatible version and a specific error messages
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -343,7 +297,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: empty
             // Reg: 9999.3.57, 9999.3.4-dummy
             // Expected: no compatible version and a specific error message
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -364,7 +318,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.3.300, 9999.7.304-global-dummy
             // Reg: 9999.3.57, 9999.3.4-dummy
             // Expected: no compatible version and a specific error message
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -385,7 +339,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.3.300, 9999.7.304-global-dummy
             // Reg: 9999.3.57, 9999.3.4-dummy, 9999.3.304
             // Expected: 9999.3.304 from reg dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -406,7 +360,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.3.300, 9999.7.304-global-dummy, 9999.3.399, 9999.3.399-dummy, 9999.3.400
             // Reg: 9999.3.57, 9999.3.4-dummy, 9999.3.304
             // Expected: 9999.3.399 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -428,7 +382,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.3.300, 9999.7.304-global-dummy, 9999.3.399, 9999.3.399-dummy, 9999.3.400, 9999.3.2400, 9999.3.3004
             // Reg: 9999.3.57, 9999.3.4-dummy, 9999.3.304, 9999.3.2400, 9999.3.3004
             // Expected: 9999.3.399 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -449,7 +403,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.3.300, 9999.7.304-global-dummy, 9999.3.399, 9999.3.399-dummy, 9999.3.400, 9999.3.2400, 9999.3.3004
             // Reg: 9999.3.57, 9999.3.4-dummy, 9999.3.304, 9999.3.2400, 9999.3.3004, 9999.3.304-global-dummy
             // Expected: 9999.3.304-global-dummy from reg dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -462,7 +416,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 .And.HaveStdErrContaining(Path.Combine(_regSelectedMessage, "9999.3.304-global-dummy", _dotnetSdkDllMessageTerminator));
 
             // Verify we have the expected SDK versions
-            dotnet.Exec("--list-sdks")
+            DotNet.Exec("--list-sdks")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -493,11 +447,6 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 return;
             }
 
-            var fixture = PortableAppFixture
-                .Copy();
-
-            var dotnet = fixture.BuiltDotnet;
-
             // Add SDK versions
             AddAvailableSdkVersions(_regSdkBaseDir, "9999.0.4");
 
@@ -507,7 +456,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: empty
             // Reg: 9999.0.4
             // Expected: 9999.0.4 from reg dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -528,7 +477,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.0.4
             // Reg: 9999.0.4
             // Expected: 9999.0.4 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -561,15 +510,9 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 return;
             }
 
-            var fixture = PortableAppFixture
-                .Copy();
-
-            var dotnet = fixture.BuiltDotnet;
-
             using (var regKeyOverride = new RegisteredInstallKeyOverride())
             {
-                string architecture = fixture.CurrentRid.Split('-')[1];
-                regKeyOverride.SetInstallLocation(_regDir, architecture);
+                regKeyOverride.SetInstallLocation(_regDir, RepoDirectories.BuildArchitecture);
 
                 // Add SDK versions
                 AddAvailableSdkVersions(_regSdkBaseDir, "9999.0.4");
@@ -580,7 +523,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 // Exe: empty
                 // Reg: 9999.0.4
                 // Expected: 9999.0.4 from reg dir
-                dotnet.Exec("help")
+                DotNet.Exec("help")
                     .WorkingDirectory(_currentWorkingDir)
                     .WithUserProfile(_userDir)
                     .Environment(s_DefaultEnvironment)
@@ -603,11 +546,6 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 return;
             }
 
-            var fixture = PortableAppFixture
-                .Copy();
-
-            var dotnet = fixture.BuiltDotnet;
-
             // Add SDK versions
             AddAvailableSdkVersions(_regSdkBaseDir, "9999.0.0", "9999.0.3-dummy");
 
@@ -617,7 +555,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: empty
             // Reg: 9999.0.0, 9999.0.3-dummy
             // Expected: 9999.0.3-dummy from reg dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -638,7 +576,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.0.3
             // Reg: 9999.0.0, 9999.0.3-dummy
             // Expected: 9999.0.3 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -661,7 +599,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.0.3
             // Reg: 9999.0.0, 9999.0.3-dummy, 9999.0.100
             // Expected: 9999.0.100 from reg dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -682,7 +620,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.0.3, 9999.0.80
             // Reg: 9999.0.0, 9999.0.3-dummy, 9999.0.100
             // Expected: 9999.0.100 from reg dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -703,7 +641,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.0.3, 9999.0.80, 9999.0.5500000
             // Reg: 9999.0.0, 9999.0.3-dummy, 9999.0.100
             // Expected: 9999.0.5500000 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -724,7 +662,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Exe: 9999.0.3, 9999.0.80, 9999.0.5500000
             // Reg: 9999.0.0, 9999.0.3-dummy, 9999.0.100, 9999.0.52000000
             // Expected: 9999.0.52000000 from reg dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -737,7 +675,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 .And.HaveStdErrContaining(Path.Combine(_regSelectedMessage, "9999.0.52000000", _dotnetSdkDllMessageTerminator));
 
             // Verify we have the expected SDK versions
-            dotnet.Exec("--list-sdks")
+            DotNet.Exec("--list-sdks")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -755,38 +693,30 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 .And.HaveStdOutContaining("9999.0.52000000");
         }
 
-        // This method adds a list of new sdk version folders in the specified
-        // sdkBaseDir. The files are copied from the _sdkDir. Also, the dotnet.runtimeconfig.json
-        // file is overwritten in order to use a dummy framework version (9999.0.0)
-        // Remarks:
-        // - If the sdkBaseDir does not exist, then a DirectoryNotFoundException
-        //   is thrown.
-        // - If a specified version folder already exists, then it is deleted and replaced
-        //   with the contents of the _builtSharedFxDir.
+        // This method adds a list of new sdk version folders in the specified directory.
+        // The actual contents are 'fake' and the mininum required for SDK discovery.
+        // The dotnet.runtimeconfig.json created uses a dummy framework version (9999.0.0)
         private void AddAvailableSdkVersions(string sdkBaseDir, params string[] availableVersions)
         {
-            DirectoryInfo sdkBaseDirInfo = new DirectoryInfo(sdkBaseDir);
-
-            if (!sdkBaseDirInfo.Exists)
-            {
-                throw new DirectoryNotFoundException();
-            }
-
             string dummyRuntimeConfig = Path.Combine(RepoDirectories.RepoRoot, "src", "test", "Assets", "TestUtils",
                 "SDKLookup", "dotnet.runtimeconfig.json");
 
             foreach (string version in availableVersions)
             {
                 string newSdkDir = Path.Combine(sdkBaseDir, version);
-                SharedFramework.CopyDirectory(_sdkDir, newSdkDir);
+                Directory.CreateDirectory(newSdkDir);
 
+                // ./dotnet.dll
+                File.WriteAllText(Path.Combine(newSdkDir, "dotnet.dll"), string.Empty);
+
+                // ./dotnet.runtimeconfig.json
                 string runtimeConfig = Path.Combine(newSdkDir, "dotnet.runtimeconfig.json");
                 File.Copy(dummyRuntimeConfig, runtimeConfig, true);
             }
         }
 
         // Put a global.json file in the cwd in order to specify a CLI
-        public void SetGlobalJsonVersion(string globalJsonFileName)
+        private void SetGlobalJsonVersion(string globalJsonFileName)
         {
             string destFile = Path.Combine(_currentWorkingDir, "global.json");
             string srcFile = Path.Combine(RepoDirectories.RepoRoot, "src", "test", "Assets", "TestUtils",
