+++ src/test/HostActivationTests/SDKLookup.cs
@@ -2,10 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.DotNet.Cli.Build;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using Xunit;
 
 namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
@@ -20,9 +20,10 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             {"COMPlus_ReadyToRun", "0" },
         };
 
-        private RepoDirectoriesProvider RepoDirectories;
-        private TestProjectFixture PortableAppFixture;
+        private readonly RepoDirectoriesProvider RepoDirectories;
+        private readonly DotNetCli DotNet;
 
+        private readonly string _baseDir;
         private readonly string _currentWorkingDir;
         private readonly string _userDir;
         private readonly string _executableDir;
@@ -30,35 +31,27 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
         private readonly string _userSdkBaseDir;
         private readonly string _exeSdkBaseDir;
         private readonly string _exeSelectedMessage;
-        private readonly string _sdkDir;
-        private readonly string _baseDir;
 
         private const string _dotnetSdkDllMessageTerminator = "dotnet.dll]";
 
         public SDKLookup()
         {
-            // From the artifacts dir, it's possible to find where the sharedFrameworkPublish folder is. We need
-            // to locate it because we'll copy its contents into other folders
-            string artifactsDir = Environment.GetEnvironmentVariable("TEST_ARTIFACTS");
-            string builtDotnet = Path.Combine(artifactsDir, "sharedFrameworkPublish");
-
             // The dotnetSDKLookup dir will contain some folders and files that will be
             // necessary to perform the tests
-            string baseDir = Path.Combine(artifactsDir, "dotnetSDKLookup");
+            string baseDir = Path.Combine(TestArtifact.TestArtifactsPath, "dotnetSDKLookup");
             _baseDir = SharedFramework.CalculateUniqueTestDirectory(baseDir);
 
             // The three tested locations will be the cwd, the user folder and the exe dir. cwd and user are no longer supported.
             //     All dirs will be placed inside the base folder
-
             _currentWorkingDir = Path.Combine(_baseDir, "cwd");
             _userDir = Path.Combine(_baseDir, "user");
             _executableDir = Path.Combine(_baseDir, "exe");
 
-            // It's necessary to copy the entire publish folder to the exe dir because
-            // we'll need to build from it. The CopyDirectory method automatically creates the dest dir
-            SharedFramework.CopyDirectory(builtDotnet, _executableDir);
+            DotNet = new DotNetBuilder(_baseDir, Path.Combine(TestArtifact.TestArtifactsPath, "sharedFrameworkPublish"), "exe")
+                .AddMicrosoftNETCoreAppFramework("9999.0.0")
+                .Build();
 
-            RepoDirectories = new RepoDirectoriesProvider(builtDotnet: _executableDir);
+            RepoDirectories = new RepoDirectoriesProvider(builtDotnet: DotNet.BinPath);
 
             // SdkBaseDirs contain all available version folders
             _cwdSdkBaseDir = Path.Combine(_currentWorkingDir, "sdk");
@@ -69,42 +62,14 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             Directory.CreateDirectory(_cwdSdkBaseDir);
             Directory.CreateDirectory(_userSdkBaseDir);
             Directory.CreateDirectory(_exeSdkBaseDir);
-
-            // Restore and build PortableApp from exe dir
-            PortableAppFixture = new TestProjectFixture("PortableApp", RepoDirectories)
-                .EnsureRestored(RepoDirectories.CorehostPackages)
-                .BuildProject();
-            var fixture = PortableAppFixture;
-
-            // Set a dummy framework version (9999.0.0) in the exe sharedFx location. We will
-            // always pick the framework from this to avoid interference with the sharedFxLookup
-            string exeDirDummyFxVersion = Path.Combine(_executableDir, "shared", "Microsoft.NETCore.App", "9999.0.0");
-            string builtSharedFxDir = fixture.BuiltDotnet.GreatestVersionSharedFxPath;
-            SharedFramework.CopyDirectory(builtSharedFxDir, exeDirDummyFxVersion);
-
-            // The actual SDK version can be obtained from the built fixture. We'll use it to
-            // locate the sdkDir from which we can get the files contained in the version folder
-            string sdkBaseDir = Path.Combine(fixture.SdkDotnet.BinPath, "sdk");
-
-            var sdkVersionDirs = Directory.EnumerateDirectories(sdkBaseDir)
-                .Select(p => Path.GetFileName(p));
-
-            string greatestVersionSdk = sdkVersionDirs
-                .Where(p => !string.Equals(p, "NuGetFallbackFolder", StringComparison.OrdinalIgnoreCase))
-                .OrderByDescending(p => p.ToLower())
-                .First();
-
-            _sdkDir = Path.Combine(sdkBaseDir, greatestVersionSdk);
-
+            
             // Trace messages used to identify from which folder the SDK was picked
             _exeSelectedMessage = $"Using dotnet SDK dll=[{_exeSdkBaseDir}";
         }
 
         public void Dispose()
         {
-            PortableAppFixture.Dispose();
-
-            if (!TestProject.PreserveTestRuns())
+            if (!TestArtifact.PreserveTestRuns())
             {
                 Directory.Delete(_baseDir, true);
             }
@@ -113,18 +78,13 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
         [Fact]
         public void SdkLookup_Global_Json_Single_Digit_Patch_Rollup()
         {
-            var fixture = PortableAppFixture
-                .Copy();
-
-            var dotnet = fixture.BuiltDotnet;
-
             // Set specified SDK version = 9999.3.4-global-dummy
             SetGlobalJsonVersion("SingleDigit-global.json");
 
             // Specified SDK version: 9999.3.4-global-dummy
             // Exe: empty
             // Expected: no compatible version and a specific error messages
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -144,7 +104,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.4-global-dummy
             // Exe: 9999.4.1, 9999.3.4-dummy
             // Expected: no compatible version and a specific error message
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -162,7 +122,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.4-global-dummy
             // Exe: 9999.4.1, 9999.3.4-dummy, 9999.3.3
             // Expected: no compatible version and a specific error message
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -180,7 +140,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.4-global-dummy
             // Exe: 9999.4.1, 9999.3.4-dummy, 9999.3.3, 9999.3.4
             // Expected: 9999.3.4 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -197,7 +157,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.4-global-dummy
             // Exe: 9999.4.1, 9999.3.4-dummy, 9999.3.3, 9999.3.4, 9999.3.5-dummy
             // Expected: 9999.3.5-dummy from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -214,7 +174,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.4-global-dummy
             // Exe: 9999.4.1, 9999.3.4-dummy, 9999.3.3, 9999.3.4, 9999.3.5-dummy, 9999.3.600
             // Expected: 9999.3.5-dummy from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -231,7 +191,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.4-global-dummy
             // Exe: 9999.4.1, 9999.3.4-dummy, 9999.3.3, 9999.3.4, 9999.3.5-dummy, 9999.3.600, 9999.3.4-global-dummy
             // Expected: 9999.3.4-global-dummy from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -243,7 +203,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 .And.HaveStdErrContaining(Path.Combine(_exeSelectedMessage, "9999.3.4-global-dummy", _dotnetSdkDllMessageTerminator));
 
             // Verify we have the expected SDK versions
-            dotnet.Exec("--list-sdks")
+            DotNet.Exec("--list-sdks")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -263,18 +223,13 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
         [Fact]
         public void SdkLookup_Global_Json_Two_Part_Patch_Rollup()
         {
-            var fixture = PortableAppFixture
-                .Copy();
-
-            var dotnet = fixture.BuiltDotnet;
-
             // Set specified SDK version = 9999.3.304-global-dummy
             SetGlobalJsonVersion("TwoPart-global.json");
 
             // Specified SDK version: 9999.3.304-global-dummy
             // Exe: empty
             // Expected: no compatible version and a specific error messages
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -292,7 +247,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.304-global-dummy
             // Exe: 9999.3.57, 9999.3.4-dummy
             // Expected: no compatible version and a specific error message
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -310,7 +265,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.304-global-dummy
             // Exe: 9999.3.57, 9999.3.4-dummy, 9999.3.300, 9999.7.304-global-dummy
             // Expected: no compatible version and a specific error message
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -328,7 +283,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.304-global-dummy
             // Exe: 99999.3.57, 9999.3.4-dummy, 9999.3.300, 9999.7.304-global-dummy, 9999.3.304
             // Expected: 9999.3.304 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -345,7 +300,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.304-global-dummy
             // Exe: 9999.3.57, 9999.3.4-dummy, 9999.3.300, 9999.7.304-global-dummy, 9999.3.304, 9999.3.399, 9999.3.399-dummy, 9999.3.400
             // Expected: 9999.3.399 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -362,7 +317,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.304-global-dummy
             // Exe: 9999.3.57, 9999.3.4-dummy, 9999.3.300, 9999.7.304-global-dummy, 9999.3.304, 9999.3.399, 9999.3.399-dummy, 9999.3.400, 9999.3.2400, 9999.3.3004
             // Expected: 9999.3.399 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -379,7 +334,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: 9999.3.304-global-dummy
             // Exe: 9999.3.57, 9999.3.4-dummy, 9999.3.300, 9999.7.304-global-dummy, 9999.3.304, 9999.3.399, 9999.3.399-dummy, 9999.3.400, 9999.3.2400, 9999.3.3004, 9999.3.304-global-dummy
             // Expected: 9999.3.304-global-dummy from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -391,7 +346,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 .And.HaveStdErrContaining(Path.Combine(_exeSelectedMessage, "9999.3.304-global-dummy", _dotnetSdkDllMessageTerminator));
 
             // Verify we have the expected SDK versions
-            dotnet.Exec("--list-sdks")
+            DotNet.Exec("--list-sdks")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -415,18 +370,13 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
         [Fact]
         public void SdkLookup_Negative_Version()
         {
-            var fixture = PortableAppFixture
-                .Copy();
-
-            var dotnet = fixture.BuiltDotnet;
-
             // Add a negative SDK version
             AddAvailableSdkVersions(_exeSdkBaseDir, "-1.-1.-1");
 
             // Specified SDK version: none
             // Exe: -1.-1.-1
             // Expected: no compatible version and a specific error messages
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -444,7 +394,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // Specified SDK version: none
             // Exe: -1.-1.-1, 9999.0.4
             // Expected: 9999.0.4 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -456,7 +406,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 .And.HaveStdErrContaining(Path.Combine(_exeSelectedMessage, "9999.0.4", _dotnetSdkDllMessageTerminator));
 
             // Verify we have the expected SDK versions
-            dotnet.Exec("--list-sdks")
+            DotNet.Exec("--list-sdks")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -470,11 +420,6 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
         [Fact]
         public void SdkLookup_Must_Pick_The_Highest_Semantic_Version()
         {
-            var fixture = PortableAppFixture
-                .Copy();
-
-            var dotnet = fixture.BuiltDotnet;
-
             // Add SDK versions
             AddAvailableSdkVersions(_exeSdkBaseDir, "9999.0.0", "9999.0.3-dummy.9", "9999.0.3-dummy.10");
 
@@ -483,7 +428,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // User: empty
             // Exe: 9999.0.0, 9999.0.3-dummy.9, 9999.0.3-dummy.10
             // Expected: 9999.0.3-dummy.10 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -502,7 +447,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // User: empty
             // Exe: 9999.0.0, 9999.0.3-dummy.9, 9999.0.3-dummy.10, 9999.0.3
             // Expected: 9999.0.3 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -523,7 +468,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // User: 9999.0.200               --> should not be picked
             // Exe: 9999.0.0, 9999.0.3-dummy.9, 9999.0.3-dummy.10, 9999.0.3, 9999.0.100
             // Expected: 9999.0.100 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -542,7 +487,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // User: 9999.0.200               --> should not be picked
             // Exe: 9999.0.0, 9999.0.3-dummy.9, 9999.0.3-dummy.10, 9999.0.3, 9999.0.100, 9999.0.80
             // Expected: 9999.0.100 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -561,7 +506,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // User: 9999.0.200               --> should not be picked
             // Exe: 9999.0.0, 9999.0.3-dummy.9, 9999.0.3-dummy.10, 9999.0.3, 9999.0.100, 9999.0.80, 9999.0.5500000
             // Expected: 9999.0.5500000 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -580,7 +525,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
             // User: 9999.0.200               --> should not be picked
             // Exe: 9999.0.0, 9999.0.3-dummy.9, 9999.0.3-dummy.10, 9999.0.3, 9999.0.100, 9999.0.80, 9999.0.5500000, 9999.0.52000000
             // Expected: 9999.0.52000000 from exe dir
-            dotnet.Exec("help")
+            DotNet.Exec("help")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -592,7 +537,7 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 .And.HaveStdErrContaining(Path.Combine(_exeSelectedMessage, "9999.0.52000000", _dotnetSdkDllMessageTerminator));
 
             // Verify we have the expected SDK versions
-            dotnet.Exec("--list-sdks")
+            DotNet.Exec("--list-sdks")
                 .WorkingDirectory(_currentWorkingDir)
                 .WithUserProfile(_userDir)
                 .Environment(s_DefaultEnvironment)
@@ -610,38 +555,30 @@ namespace Microsoft.DotNet.CoreSetup.Test.HostActivation
                 .And.HaveStdOutContaining("9999.0.52000000");
         }
 
-        // This method adds a list of new sdk version folders in the specified
-        // sdkBaseDir. The files are copied from the _sdkDir. Also, the dotnet.runtimeconfig.json
-        // file is overwritten in order to use a dummy framework version (9999.0.0)
-        // Remarks:
-        // - If the sdkBaseDir does not exist, then a DirectoryNotFoundException
-        //   is thrown.
-        // - If a specified version folder already exists, then it is deleted and replaced
-        //   with the contents of the _builtSharedFxDir.
+        // This method adds a list of new sdk version folders in the specified directory.
+        // The actual contents are 'fake' and the mininum required for SDK discovery.
+        // The dotnet.runtimeconfig.json created uses a dummy framework version (9999.0.0)
         private void AddAvailableSdkVersions(string sdkBaseDir, params string[] availableVersions)
         {
-            DirectoryInfo sdkBaseDirInfo = new DirectoryInfo(sdkBaseDir);
-
-            if (!sdkBaseDirInfo.Exists)
-            {
-                throw new DirectoryNotFoundException();
-            }
-
             string dummyRuntimeConfig = Path.Combine(RepoDirectories.RepoRoot, "src", "test", "Assets", "TestUtils",
                 "SDKLookup", "dotnet.runtimeconfig.json");
 
             foreach (string version in availableVersions)
             {
                 string newSdkDir = Path.Combine(sdkBaseDir, version);
-                SharedFramework.CopyDirectory(_sdkDir, newSdkDir);
+                Directory.CreateDirectory(newSdkDir);
 
+                // ./dotnet.dll
+                File.WriteAllText(Path.Combine(newSdkDir, "dotnet.dll"), string.Empty);
+
+                // ./dotnet.runtimeconfig.json
                 string runtimeConfig = Path.Combine(newSdkDir, "dotnet.runtimeconfig.json");
                 File.Copy(dummyRuntimeConfig, runtimeConfig, true);
             }
         }
 
         // Put a global.json file in the cwd in order to specify a CLI
-        public void SetGlobalJsonVersion(string globalJsonFileName)
+        private void SetGlobalJsonVersion(string globalJsonFileName)
         {
             string destFile = Path.Combine(_currentWorkingDir, "global.json");
             string srcFile = Path.Combine(RepoDirectories.RepoRoot, "src", "test", "Assets", "TestUtils",
